backend/requirements.txt
fastapi
uvicorn
python-dotenv
authlib
python-jose
pymongo
